refactor(regeneration): extract shared selection primitives into core/regeneration (#9928)
* refactor(regeneration): add core/regeneration package with shared models

Start a leaf package that holds the regeneration selection primitives shared by
the proposed-change pipeline and the merge follow-up. This first block carries
the models: the definition-selection flag, the impact scope and its result, the
predicate outcome, and the definition protocol the predicates read.


* refactor(regeneration): add selection predicates

Move the regeneration predicate family into the package: the query-changed and
definition-changed diff checks, the transform dependency-closure check and its
repository-diff lookup helper, the query-relevant node-change filter, and the
triggering-action helpers. Names crossing the package boundary are public.


* refactor(regeneration): add member selection helpers

Move the per-member selection helpers into the package: the subscriber-by-member
mapping, the artifact render decision, and the generator execution decision.


* refactor(regeneration): add artifact definition gather and parser

Move the artifact-definition gather query and its edge parser into the package.


* refactor(regeneration): add field-level impact resolution

Move the field-level impacted-subscriber resolution into the package: it maps a
branch's data changes onto the subscribers a query actually reads, plus the
subscriber-by-node gather query and its lookup helper.


* refactor(regeneration): source selection primitives from core.regeneration

Delete the relocated definitions from the proposed-change tasks module and
import them from core.regeneration instead, and re-point the merge
selective-regeneration package and the affected unit tests at the same source.
The merge package no longer reaches into proposed-change internals; behaviour is
unchanged.


* refactor(merge): promote post-merge dispatcher imports to top level

With the selection primitives sourced from the core.regeneration leaf package,
the merge selective-regeneration chain no longer imports the proposed-change
tasks module, so the branch-merge post-process no longer needs function-local
imports to avoid a cycle. Promote them to module scope and record the
extraction status in the follow-up specs.


* refactor(regeneration): relocate selection tests to mirror the package

Move the predicate and member-selection unit tests to
tests/unit/core/regeneration/ so the tests sit beside the production code they
exercise. The relevant-node-change test is renamed to match its predicate.
Pure relocation; the branch-diff test stays with proposed_change.


---------
